Move PluginBase into own file
